All Enemies

+++ Assets/Scripts/Fish.cs
@@ -3,8 +3,8 @@ using System;
 
 public class Fish : MonoBehaviour, IItem
 {
-    public static event Action<int> OnFishCollect;
-    public int worth = 5;
+    public static event Action<float> OnFishCollect;
+    public float worth = 5;
     private bool collected;
 
     public void Collect()
@@ -13,7 +13,7 @@ public class Fish : MonoBehaviour, IItem
         {
             return;
         }
-
+        
         collected = true;
         OnFishCollect.Invoke(worth);
         Destroy(gameObject);
